+++ CHANGELOG.md
@@ -65,6 +65,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed auto-rs-discard-granularity staying set on devices that don't support discards (e.g. LUKS without AllowDiscards)
 - Throw a CriticalError(restart) if a JDK spawn helper exception was caught (usually a JDK version mismatch)
 - Fixed bug where taking a snapshot on a not-only-DRBD resource could hang indefinitely
+- Attempt to fix rare bug causing "shipping in progress" that can only be cleared by restarting the controller
 
 ## [1.33.3] - 2026-05-11